[CHORE] Specs for Topic Messages component
